Add pipeline to trigger builds for updating base images (#194)
Defines a pipeline that invokes the rebuildStaleImages command of Image Builder. This command will read the subscriptions.json file included in these changes to determine the images that are to be checked to ensure their base images are up-to-date with the base images defined in the image info file from the versions repo. If not, Image Builder will kick off the pipeline defined in the subscriptions file so that the images get updated.